Refactor: pixelated image rendering (2)
- fix circular dependency (introduced in previous commit)
- move css constants into `src/constants/pixelated_image`
- move support function into `src/lib/supports_pixelated_image`
