+++ src/lib/index.js
@@ -1426,59 +1426,3 @@ lib.getPositionFromD3Event = function() {
         ];
     }
 };
-
-// Pixelated image rendering
-// The actual CSS declaration is prepended with fallbacks for older browsers.
-// NB. IE's `-ms-interpolation-mode` works only with <img> not with SVG <image>
-// https://developer.mozilla.org/en-US/docs/Web/CSS/image-rendering
-// https://caniuse.com/?search=image-rendering
-// http://phrogz.net/tmp/canvas_image_zoom.html
-var PIXELATED_IMAGE_DECLARATIONS = [
-    ['image-rendering', 'optimizeSpeed'],
-    ['image-rendering', '-moz-crisp-edges'],
-    ['image-rendering', '-o-crisp-edges'],
-    ['image-rendering', '-webkit-optimize-contrast'],
-    ['image-rendering', 'optimize-contrast'],
-    ['image-rendering', 'crisp-edges'],
-    ['image-rendering', 'pixelated']
-];
-
-lib.PIXELATED_IMAGE_STYLE = PIXELATED_IMAGE_DECLARATIONS.map(function(d) {
-    return d.join(': ') + '; ';
-}).join('');
-
-var _supportsPixelated = null;
-
-/**
- * Check browser support for pixelated image rendering
- *
- * @return {boolean}
- */
-lib.supportsPixelatedImage = function() {
-    if(_supportsPixelated !== null) { // only run the feature detection once
-        return _supportsPixelated;
-    }
-    if(lib.isIE()) {
-        _supportsPixelated = false;
-    } else {
-        var declarations = Array.from(PIXELATED_IMAGE_DECLARATIONS).reverse();
-        var supports = window.CSS && window.CSS.supports || window.supportsCSS;
-        if(typeof supports === 'function') {
-            _supportsPixelated = declarations.some(function(d) {
-                return supports.apply(null, d);
-            });
-        } else {
-            var Drawing = require('../components/drawing');
-            var image3 = Drawing.tester.append('image');
-            var cStyles = window.getComputedStyle(image3.node());
-            image3.attr('style', lib.PIXELATED_IMAGE_STYLE);
-            _supportsPixelated = declarations.some(function(d) {
-                var value = d[1];
-                return cStyles.imageRendering === value ||
-                       cStyles.imageRendering === value.toLowerCase();
-            });
-            image3.remove();
-        }
-    }
-    return _supportsPixelated;
-};
